+++ src/Journal/journal.ts
@@ -1,23 +1,9 @@
-import { JournalInterface, createJournal } from './journalInterface';
+import { JournalInterface, createJournal, validJournalInterface } from './journalInterface';
 import journalInfoCitation, { noPublishingInfo } from './journalInfo';
 import { getRange } from './pageRange';
 import { JournalInfoInterface } from './journalInfoInterface';
 import { BookInterface } from '../Book/bookInterface';
 import validatePages from '../noteUtility';
-
-export function journalNote(info: JournalInterface, page: {page: string}): string {
-  let citation = createJournal(info);
-  return `${citation.authorList.note()}"${citation.title}," ${journalInfoCitation(citation.info)}${isSpaceRequired(citation.info)}${page.page}.`;
-}
-
-export function journalShortNote(info: JournalInterface, page: {page: string}): string {
-  return `${getShortNotePrefix(info) + page.page}.`;
-}
-
-export function journalBibliography(info: JournalInterface): string {
-  let citation = createJournal(info);
-  return `${citation.authorList.bibliography()}"${citation.title}." ${journalInfoCitation(citation.info)} ${getRange(citation.startRange, citation.endRange)}${isPeriodRequired(citation.info, citation.startRange)}`.trim();
-}
 
 export function journalNoteList(info: JournalInterface, pages: {page: string}[]): string[] {
   const validPages: Array<{page: string}> = validatePages(pages);
@@ -34,23 +20,36 @@ export function journalNoteList(info: JournalInterface, pages: {page: string}[])
   return notes;
 }
 
-export function eJournalNote(info: JournalInterface, page: {page: string}, url: string): string {
-  return (`${(journalNote(info, page)).replace(/.$/, ', ') + url}.`);
-}
+export function journalBibliography(info: JournalInterface): string {
+  let citation = createJournal(info);
 
-export function eJournalNoteList(info: JournalInterface, pages: {page: string}[], url: string): string[] {
-  const validPages: Array<{page: string}> = validatePages(pages);
-  let notes: string[] = [];
-
-  for (let page of validPages) {
-    notes.push(eJournalNote(info, page, url));
+  if (citation.url !== '') {
+    return (`${standardJournalBibliography(citation)} ${citation.url}.`);
   }
 
-  return notes;
+  return standardJournalBibliography(citation);
 }
 
-export function eJournalBibliography(info: JournalInterface, url: string): string {
-  return (`${journalBibliography(info)} ${url}.`);
+export function journalShortNote(info: JournalInterface, page: {page: string}): string {
+  return `${getShortNotePrefix(info) + page.page}.`;
+}
+
+export function journalNote(info: JournalInterface, page: {page: string}): string {
+  let citation: validJournalInterface = createJournal(info);
+
+  if (citation.url !== '') {
+    return (`${(standardJournalNote(citation, page)).replace(/.$/, ', ') + citation.url}.`);
+  }
+
+  return standardJournalNote(citation, page);
+}
+
+function standardJournalNote(citation: validJournalInterface, page: {page: string}): string {
+  return `${citation.authorList.note()}"${citation.title}," ${journalInfoCitation(citation.info)}${isSpaceRequired(citation.info)}${page.page}.`;
+}
+
+function standardJournalBibliography(citation: validJournalInterface): string {
+  return `${citation.authorList.bibliography()}"${citation.title}." ${journalInfoCitation(citation.info)} ${getRange(citation.startRange, citation.endRange)}${isPeriodRequired(citation.info, citation.startRange)}`.trim();
 }
 
 function isSpaceRequired(info: JournalInfoInterface): string {