+++ src/cmosCitation.ts
@@ -5,9 +5,7 @@ import { LectureInterface } from './Lecture/lectureInterface';
 import { bookBibliography, bookNoteList } from './Book/book';
 import { webpageBibliography, webpageNote } from './Website/webpage';
 import { lectureBibliography, lectureNote } from './Lecture/lecture';
-import {
-  eJournalBibliography, eJournalNoteList, journalBibliography, journalNoteList,
-} from './Journal/journal';
+import { journalBibliography, journalNoteList } from './Journal/journal';
 
 export function book(info: BookInterface, pages: string[]): {bibliography: string, notes: string[]} {
   return { bibliography: bookBibliography(info), notes: bookNoteList(info, pages) };
@@ -15,11 +13,6 @@ export function book(info: BookInterface, pages: string[]): {bibliography: strin
 
 export function journal(info: JournalInterface, pages: {page: string}[]): {bibliography: string, notes: string[]} {
   return { bibliography: journalBibliography(info), notes: journalNoteList(info, pages) };
-}
-
-export function eJournal(info: JournalInterface, pages: {page: string}[]): {bibliography: string, notes: string[]} {
-  if (!info.url) throw new Error('URL not specified');
-  return { bibliography: eJournalBibliography(info, info.url), notes: eJournalNoteList(info, pages, info.url) };
 }
 
 export function webpage(info: WebpageInterface): {bibliography: string, notes: string[]} {